+++ src/files.c
@@ -22,25 +22,11 @@
 
 #include <stdlib.h>
 
-static ino_t current_ino = PFS_FILE_MIN_USED_INO;
+static ino_t current_ino = PFS_FILE_INO_MIN;
 
-/*
-Get next inode number between PFS_FILE_MIN_USED_INO and PFS_FILE_MAX_USED_INO, inclusive.
-*/
-static ino_t next_ino () {
-	if (current_ino > PFS_FILE_MAX_USED_INO)
-		return current_ino;
-	return current_ino++;
-}
-
-/*
-Create a new pfs_file.
-@parameter full_path: The path of the file
-@parameter mode: The mode of the file
-*/
 pfs_file* pfs_file_create (char* full_path, mode_t mode) {
-	ino_t new_ino = next_ino ();
-	if (new_ino == PFS_FILE_MAX_USED_INO)
+	ino_t new_ino = pfs_file_next_ino ();
+	if (new_ino == 0)
 		return NULL;
 
 	pfs_file* file = malloc (sizeof (*file));
@@ -57,19 +43,21 @@ pfs_file* pfs_file_create (char* full_path, mode_t mode) {
 	return file;
 }
 
-/*
-Free a pfs_file. Should only be called if deleted from the file table.
-@parameter file: The pfs_file to free
-*/
 void pfs_file_free (pfs_file* file) {
 	g_string_free (file->path, TRUE);
 	free (file);
 }
 
-/*
-Same as pfs_file_free, but for use with GHashTable.
-@parameter file: The pfs_file to free
-*/
 void pfs_file_free_void (void* file) {
 	pfs_file_free ((pfs_file*)file);
+}
+
+ino_t pfs_file_next_ino (void) {
+	if (current_ino > PFS_FILE_INO_MAX)
+		return 0;
+	return current_ino++;
+}
+
+fsfilcnt_t pfs_file_used_ino_count (void) {
+	return (fsfilcnt_t)(current_ino - PFS_FILE_INO_MIN);
 }