+++ src/files.h
@@ -24,6 +24,8 @@
 
 #include <glib.h>
 #include <sys/stat.h>
+#include <sys/statvfs.h>
+#include <sys/types.h>
 
 typedef struct {
 	GString* path;
@@ -32,21 +34,45 @@ typedef struct {
 	ino_t ino;
 } pfs_file;
 
+/*
+Create a new pfs_file.
+@parameter full_path: The path of the file
+@parameter mode: The mode of the file
+*/
 pfs_file* pfs_file_create (char* path, mode_t mode);
+
+/*
+Free a pfs_file. Should only be called if deleted from the file table.
+@parameter file: The pfs_file to free
+*/
 void pfs_file_free (pfs_file*);
+
+/*
+Same as pfs_file_free, but for use with GHashTable.
+@parameter file: The pfs_file to free
+*/
 void pfs_file_free_void (void*);
 
 /*
-This is the minimum inode number used by pfs_file_create();
-Smaller values can be used freely for other things.
+Get next inode number between PFS_FILE_INO_MIN and PFS_FILE_INO_MAX, inclusive.
+Returns 0 if the pool is exhausted (quite unlikely to happen in reality).
 */
-#define PFS_FILE_MIN_USED_INO ((ino_t)1 << 16)
+ino_t pfs_file_next_ino (void);
+
+/*
+Get count of currently used inode numbers.
+Numbers are never returned, so deleting files does not decrease this number.
+*/
+fsfilcnt_t pfs_file_used_ino_count (void);
+
+/*
+This is the minimum inode number used by pfs_file_next_ino() and pfs_file_create().
+Smaller values can be used for other things, if required.
+*/
+#define PFS_FILE_INO_MIN ((ino_t)1 << 16)
 /*
 And this is the maximum.
-Most files that can be created is
-  (PFS_FILE_MAX_USED_INO - PFS_FILE_MIN_USED_INO + 1),
-though, realistically, memory will run out first.
 */
-#define PFS_FILE_MAX_USED_INO ((ino_t)1 << (sizeof(ino_t) * 8 - 2))
+#define PFS_FILE_INO_MAX ((ino_t)1 << (sizeof(ino_t) * 8 - 2))
 
 #endif // PLAYLISTFS_FILES_H